+++ packages/deploy/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @openfn/deploy
 
+## 0.15.0
+
+### Minor Changes
+
+- b5cf7b5: Support custom webhook paths in project.yaml
+
 ## 0.14.0
 
 ### Minor Changes